Refactor MeasurementSet into MeasurementSet module
Having one class per module seems to work well, and I intend to extend
this to other modules.
